extraction of writer complete
