+++ cli/htmlWriter.js
@@ -4,13 +4,15 @@
 var fs = require("fs");
 var html = require('html');
 var shellTools = require('./shellTools');
+var genericWriter = require('./genericWriter');
 
+//might be extractable
 function generateFile(directory, outputFile) {
-	if (!fs.existsSync(directory)) {
-        throw new URIError(directory + ' is not a directory');
-    }
-    var porRepo = getPORObjectFromRepo(directory, false);
-    return writePORObjectToHTMLFile(porRepo, outputFile);
+    return genericWriter.generateFile(directory, outputFile, this.writePORObjectToHTMLFile)
+}
+
+function getPORObjectFromRepo(currentDir, textHaveParents){
+    return genericWriter.getPORObjectFromRepo(currentDir, textHaveParents);
 }
 
 function writePORObjectToHTMLFile(porObject, outputFile) {
@@ -18,48 +20,6 @@ function writePORObjectToHTMLFile(porObject, outputFile) {
 
     fs.writeFileSync(outputFile, fileString);
     return fileString;
-}
-
-/**
-* Returns a POR object of the repo.
-*/
-function getPORObjectFromRepo(currentDir, textHaveParents){
-
-	var metadataLocation = currentDir + "/metadata.json";
-	var metadata = JSON.parse(fs.readFileSync(metadataLocation));
-    var objectId = currentDir.replace(/^.*[\\\/]/, '');
-
-    var porObject = {
-        porID: objectId,
-        metadata: metadata,
-        children: []
-    };
-
-    // Recursively builds the por object from the construction order.
-    metadata.constructionOrder.forEach(function(id) {
-        var path = currentDir + "/" + id;
-        if (fs.existsSync(path)) {
-            // For directories.
-            if (fs.lstatSync(path).isDirectory()){
-                porObject.children.push(getPORObjectFromRepo(path, textHaveParents));
-            }
-        } else {
-            // For files, currently only text are used.
-            var fileLoc = path + ".txt";
-            if (fs.lstatSync(fileLoc).isFile() && fs.existsSync(fileLoc)){
-                var textValue = {
-                    value: fs.readFileSync(fileLoc, "utf-8"),
-                    porID: fileLoc.replace(/^.*[\\\/]/, '').split('.')[0]
-                };
-                if(textHaveParents) {
-                    textValue.parent = objectId;
-                }
-                porObject.children.push(textValue);
-            }
-        }
-    });
-
-    return porObject;
 }
 
 function convertPORRepoObjectToHTMLString(porObject) {
@@ -85,20 +45,17 @@ function recursivelyConvertPORObjectToHTML(porObject){
 			});
 		}
 	} else {
-        fileString += convertTextNodeToHTMLString(porObject);
+        fileString += porObject.value;
     }
 
     return fileString;
 }
 
+//This remains here for now for future extensability.
 function convertTextNodeToHTMLString(porObject) {
     // if the node is a leaf (text file)
     var objectString = "";
-
-    //We don't want to tag text nodes, as this might break something that uses the
-    //HTML, so just give the text
     objectString += porObject.value;
-
     return objectString;
 }
 
@@ -112,7 +69,6 @@ function convertTagNodeToHTMLString(porObject) {
         objectString += recursivelyConvertPORObjectToHTML(child);
     });
 
-
     // End tag
     if (emptyTags.indexOf(porObject.metadata.tag) == -1) {
         objectString += "</" + porObject.metadata.tag + ">";
@@ -123,7 +79,7 @@ function convertTagNodeToHTMLString(porObject) {
 function extractOpeningTag(porObject) {
     var objectString = "<" + porObject.metadata.tag;
     porObject.metadata.attributes.forEach(function(attribute) {
-        // This is wrong and temporary, we need to revise the initial parsing so that we don't make everything strings.
+        // This is wrong and 'temporary', we need to revise the initial parsing so that we don't make everything strings.
         // TODO: Change repository creation to keep data types
         objectString += ' ' + attribute.name + '="' + attribute.value + '"';
     });
@@ -131,14 +87,9 @@ function extractOpeningTag(porObject) {
     return objectString;
 }
 
+//no html
 function getPreviousFileVersions(repoLocation) {
-    shellTools.checkoutToCommit(repoLocation, 'HEAD^');
-    var oldObject = getPORObjectFromRepo(repoLocation, true);
-
-    shellTools.checkoutToCommit(repoLocation, 'master');
-    var newObject = getPORObjectFromRepo(repoLocation, true);
-
-    return [oldObject, newObject];
+    return genericWriter.getPreviousFileVersions(repoLocation);
 }
